requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via pynacl
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via typer
colorama==0.4.6
    # via typer
commonmark==0.9.1
    # via rich
deprecated==1.2.13
    # via pygithub
idna==3.4
    # via requests
loguru==0.6.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygithub==1.57
    # via -r requirements.in
pygments==2.13.0
    # via rich
pyjwt==2.6.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.28.1
    # via pygithub
rich==12.6.0
    # via typer
shellingham==1.5.0
    # via typer
typer[all]==0.7.0
    # via -r requirements.in
urllib3==1.26.13
    # via requests
wrapt==1.14.1
    # via deprecated
